refactor: proxy write
