+++ proxy/src/http/prom.rs
@@ -65,15 +65,13 @@ impl<Q: QueryExecutor + 'static> Proxy<Q> {
         };
         // TODO: Define a new public method instead of calling the grpc write method
         // directly.
-        let result = self.handle_write(ctx, table_request).await;
-        if let Some(header) = result.header {
-            if header.code != StatusCode::OK.as_u16() as u32 {
-                ErrNoCause {
-                    code: StatusCode::INTERNAL_SERVER_ERROR,
-                    msg: format!("fail to write storage, err:{:?}", header.error),
-                }
-                .fail()?;
+        let result = self.handle_write_internal(ctx, table_request).await?;
+        if result.failed != 0 {
+            ErrNoCause {
+                code: StatusCode::INTERNAL_SERVER_ERROR,
+                msg: format!("fail to write storage, failed rows:{:?}", result.failed),
             }
+            .fail()?;
         }
 
         Ok(())
